Expose SGroup data label API in installed header
